Refactored action menu
